requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
binance.py==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coinbase==2.1.0
constantly==23.10.4
cryptography==45.0.5
dateparser==1.2.2
dnspython==2.7.0
email-validator==2.0.0
fastapi==0.104.1
frozenlist==1.7.0
h11==0.16.0
http_client==0.1.22
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.2
loguru==0.7.0
MarkupSafe==3.0.2
multidict==6.6.3
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pycurl==7.45.6
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyOpenSSL==25.1.0
python-binance==0.7.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
regex==2024.11.6
requests==2.31.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.23.2
websockets==15.0.1
yarl==1.20.1
zope.interface==7.2
